feat: refactor cli internals, add langgraph runtime factory
